+++ packages/beff-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @beff/client
 
+## 0.0.137
+
+### Patch Changes
+
+- more improvements to union errros
+- Updated dependencies
+  - @beff/cli@0.0.137
+
 ## 0.0.136
 
 ### Patch Changes